feat(authentication): implement support to retrieve extra claims
k8s' current implementation of OIDC only retrieve the username and groups from
the token however it contains more than that and depending on how the client is
configured, extra useful information can be present. The downside of the current
implementation (the kube-apiserver is the same) is that there's is no way to get
these extra fields.

This patch implements an additional field to configure a list of extra claims
that should be extracted from the token and loaded into the UserInfo Extra field
for future processing.