Do not automatically set audiences instead of clientID
Such modification can be a disruptive change for users, e.g. an integration can rely on reading the clientID from an oidc object. If we unset it this can cause stuff to break or act in an unpredictable manner. We leave users to perform the migration on their own terms.